+++ test/cookbooks/nfs_test/recipes/default.rb
@@ -1,10 +1,6 @@
-# These services on centos/fedora are loaded lazily and wait until a
-# connection is attempted to start so, manually start them here so that
-# Kitchen can test for them
+# This service is loaded lazily and wait until a connection is attempted to
+# start so, manually start them here so that Kitchen can test for them
 
-%w(rpcbind rpc-statd).each do |s|
-  service s do
-    action [:enable, :start]
-    only_if { platform_family?('rhel', 'fedora') }
-  end
+service 'rpcbind' do
+  action [:enable, :start]
 end