+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This file is used to list changes made in each version of the nfs cookbook.
 - Loosen version pin on line cookbook
 - Fix CentOS 8+ and Fedora and properly manage /etc/nfs.conf
 - Add `fsid` property to the `nfs_export` resource
+- Fix services that are loaded
 
 ## 4.0.0 - *2021-09-11*